requirements.txt
# 核心依赖
ccxt>=4.0.0
pandas>=2.0.0
pandas-ta>=0.3.14b
numpy>=1.24.0
asyncio-compat>=0.1.0

# 数值计算
decimal>=1.70

# 配置管理
python-dotenv>=1.0.0
pydantic>=2.0.0

# 日志处理
structlog>=23.0.0
rich>=13.0.0

# HTTP客户端
aiohttp>=3.8.0
requests>=2.31.0

# 时间处理
python-dateutil>=2.8.0
pytz>=2023.3

# 数据验证
marshmallow>=3.19.0

# 开发工具
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# 可选依赖
jupyter>=1.0.0
matplotlib>=3.7.0
plotly>=5.15.0